1st working version
This is the first working version of the late-time lc binning program, and is made to detect re-brightening of a SN at late times.
After a list of csv files are provided, the program goes through this list, estimate the date of peak SN light, and binning the observations considered to be at late times compared to this peak.

In the event that the brightness of the peak is at / brighter than a set magnitude, a linear model is fitted to the tail in mag space, spanning from 50 to 200 days after the peak. Using this model, the tail of the SN is removed from the late-time observations, up until the model tail crosses a mag threshold. This is done to remove false positives from close, bright SN whose tails can be followed for a very long time.

For each object in the given list, the following is made:
- A plot of the light curve in every filter (+ the fitted line if it was used)
- Plots with the result of the binning procedure (1 per filter)
- A csv file with all the binning results
- A csv file with the settings and constants used, including the fitted parameters

Additionally, an overview file is made listing the most promising bins in each filter of each object