Leak closed using multiprocessing
With this version, a pool of workers is created with its size equal to the available cpu cores in order to some of the calculations in parallel. As a result of this, the whole process speeds up, and because the Matplotlib objects are dumped after a worker is finished with 1 light curve, this removes the slow-down issue due to memory clogging entirely. The result is a program that runs up to 33x faster than the old version (when it was tested on 653 light curves).

Additionally, the option not to make & save plots was added, which might be useful while testing the program using fake light curves for instance. As making & saving plots is relatively time consuming, not doing this decreased the time it took to complete the task by a factor of 3 compared to the previous attempt. This results in the program now being more than 100x as fast as it was before multiprocessing.

The parallelisation of the program resulted in the old progress bar not working, as expected. This has been fixed using the tqdm package.