Corrected version
This version has been corrected for the issues in the older version, with more permanent solutions than the temporary ones in the previous correction.
This should fix both issues #7 and #8, although the name extraction now requires '_S' to be in the filename in order for it to cut at the right place, e.g. 'ZTF18aasdted_SNT_5.0.csv'. Otherwise it will just use the entire name.
The empty data frame issue has been fixed by checking for it and stopping processing the object early if there is nothing to do. If there are less than 2 datapoints, there is no way anything can be put into bins, nor are there any objects with values smaller than the mean. --> No peak can be found, so it will be put as None, which is then picked up when deciding to continue to the binning or go to the next object.