Forgotten check
I forgot to check if host data with a matching name was found in this particular case. The other 2 did have that check, so the issue wasn't triggered before.
Spacing might be a little off, since this is a quick fix (lines 747-749), will check & fix this later.

I have found several other issues, but I'm not sure yet if it is the code or data that needs fixing. This branch will be merged after all of these are investigated & fixed.