Switching to idr file format
Changes in the file format forced some additional changes. Fixed some other issues as well in the mean time.
While this version has been used at some point, it needs some cleaning to get it to a proper final state. I will be working from this file to do so and try to integrate the version for the simulations if possible (otherwise I'll just add that version separately).