Clean-up
Clean-up of the binning program, removing a lot of the mess created when switching to the idr input file format. This might be followed up with a commit about the version of the program suited for simulations & some more / clearer comments.

Also moved the parameters intended to be changed if needed to a config file. To run the program one should now type:
python3 Binning_program.py <lc loc> <save loc> <config file>