Operation mode unification
Starting a new branch for this part of the effort to getting everything back up to date again.
Up until now I had several versions intended to be used with different types of input (DR2-like csv, FPbot-like csv, simsurvey result).

The idea is to put everything into the same code, and choosing a 'mode'. Each one has its own instruction on how to generate the list of objects to be put through the pipeline, and its own load_source() function. In the load_source() function the object should be loaded and readied for the pipeline, which includes going to a more DR2-like format, as that is what the pipeline is expecting to receive.

At the moment the simsurvey lc is transformed into an FPbot-like input, but the step from FPbot-like to DR2-like has not been implemented yet. For this I need to go through the code that does this and add the relevant parts here (code received from MatSmithAstro).

After this the modes need to be tested before these branches can be merged of course, to make sure that they all work as intended.

One major advantage of this (as well as putting stuff in the config file) is that it should be relatively easy to adapt to a new input type / science case. For instance, it should be straight forward to specify a required baseline period instead of letting the pipeline determine this, or specify peak dates instead of finding them. All that is needed is to add a new mode with its own load_source() function.