Mode combination
Last version made a start with the effort in combining the versions of the code that allowed me to use different input formats. Now the 3 main ones have been combined into 1 program, and can be selected by choosing the mode. As a bonus to this method, it is now possible to easily assume a new input type by adding a new mode and corresponding load_source() method (and whatever extra functions the new mode might need).

The modes are ZTFDR2, simulations, and fpbot. These assume input in the formats of ZTFDR2, simsurvey result pickle, and raw fpbot, respectively.

The versions have been tested and compared to each other and the older versions of the code for each type. The old and new versions give identical results (e.g. ZTFDR2 mode in this version of the code gives the same result as using the older code made for this input type)

There are however small differences between the fpbot & ZTFDR2 output. These differences however seem  too small to significantly alter the final_verdicts result. I want to confirm this by running a couple of edge-cases through the pipeline before joining this branch back in with the main one. This will be done later this week, and at that time I will also rid the code of some unneeded comments that are still scattered throughout at the moment.